Add the Quote component. Update the data type of some Bar attributes. Add an QuoteSide enum to determine if a quote is an ask or bid

+++ AlpacaTrader/Component/Bar.cs
@@ -46,7 +46,7 @@ public class Bar
     ///   The total number of shares traded in the time period.
     /// </summary>
     [JsonPropertyName("v")]
-    public long Volume { get; set; }
+    public int Volume { get; set; }
         
     /// <summary>
     ///   The total number of trades made in the time period.
